Refactored Tests, Parse List Paramter in Config
