+++ src/conftest.py
@@ -1,0 +1,1 @@
+# Allows pytest to find modules under "./src" without having to prefix imports with "src.".
