+++ src/polyswarmdconfig/config.py
@@ -9,6 +9,7 @@ from typing import Any, Dict, Tuple, Union
 
 from polyswarmdconfig.exceptions import MissingConfigValueError
 
+
 logger = logging.getLogger(__name__)
 
 
@@ -50,8 +51,10 @@ class Config:
     @classmethod
     def correct_type(cls, key: str, value: Any) -> Any:
         cast = typing.get_type_hints(cls).get(key)
-        if cast and cast in [int, str, bool]:
+        if cast and cast in [int, bool]:
             return cast(value)
+        elif cast and cast in [str]:
+            return cls._attempt_to_extract_list(value)
         else:
             return value
 
@@ -88,3 +91,11 @@ class Config:
         title = separated[0].lower()
         rest = separated[1] if len(separated) > 1 else ''
         return title, rest
+
+    @staticmethod
+    def _attempt_to_extract_list(value):
+        items_delimiter = ','
+        if items_delimiter in value:
+            return list(map(str.strip, value.split(items_delimiter)))
+        else:
+            return value
